мелкие недочеты, организация кода
