+++ src/Components/Footer/Footer.jsx
@@ -1,0 +1,16 @@
+import { Layout } from "antd";
+const { Footer } = Layout;
+
+const FooterProject = () => {
+  return (
+    <Footer
+      style={{
+        textAlign: "center",
+      }}
+    >
+      Ant Design ©2023 Created by Ant UED
+    </Footer>
+  );
+};
+
+export default FooterProject;
