+++ src/Components/FormAuthorization/FormAuthorization.jsx
@@ -1,9 +1,9 @@
 import { Input, Form, Button } from "antd";
 
 //код самой формы авторизации
-function FormAuthorization({ onFinishSignIn }) {
+function FormAuthorization({ onFinish }) {
   return (
-    <Form name="basic" onFinish={onFinishSignIn} autoComplete="off">
+    <Form name="basic" onFinish={onFinish} autoComplete="off">
       <Form.Item
         label="Почта"
         name="email"
